[신동한] DB의 파일사이즈 타입이 INT에서 BIGINT으로 변경됨에 따라, 담당 VO의 파일사이즈 타입을 long으로 변경
파일사이즈가 Byte단위이므로 int로 다 담지 못할 경우가 생길 경우를 대비하여 DB의 파일사이즈 타입을 BIGINT 타입으로
변경함

이에 따라 DB의 파일사이즈 컬럼을 받아올 수 있도록 담당한 VO(QuestionCommentFile)의 파일사이즈 타입을
long으로 변경함

+++ lms/src/main/java/gd/fintech/lms/teacher/vo/QuestionCommentFile.java
@@ -16,7 +16,7 @@ public class QuestionCommentFile {
 	private int questionCommentNo;
 	
 	// 첨부파일 용량
-	private int questionCommentFileSize;
+	private long questionCommentFileSize;
 	
 	// 첨부파일 MIME 타입
 	private String questionCommentFileType;
